Move `h.views.api_profile` -> `h.views.api.profile`
- Make test imports more consistent with other view tests
